refactor : LikesRepo 네이티브 쿼리 파라미터 설정

+++ src/main/java/com/book_everywhere/review/entity/Review.java
@@ -47,7 +47,7 @@ public class Review {
 
 
     @Transient // 칼럼이 만들어지지 않는다.
-    private int likeCount;
+    private Long likeCount;
 
     @Transient
     private boolean likeState;
